Add Broccoli Markdown
Run `broccoli serve` to preview the generated HTML files or
run `broccoli build html` to compile the files into
the `html` directory.

See https://github.com/hemanth/broccoli-md

Note: when serving the files, the `src` dir is watched and
the build is automatic, but you still need to reload the page
in your browser to see the changes.